Sorting menu components and sorting components renamed. Added PropTypes into sorting components.
